requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-APScheduler==1.13.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
python-dotenv==1.0.0
yfinance==0.2.33
gunicorn==21.2.0
requests==2.31.0
